requirements.txt
Django==5.0.1
numpy==1.26.4
opencv_python==4.8.1.78
opencv_python_headless==4.8.1.78
PyAutoGUI==0.9.54
pyserial==3.5
PyYAML==6.0.1
ultralytics==8.1.7
webdriver_manager==4.0.1
websocket_client==1.3.3
websockets==12.0
asyncio==3.4.3
onnxruntime==1.16.3
onnx==1.15.0
dill==0.3.8